Reworked the layouts/_years_bar.html to conditionally display months.
The layouts/_years_bar.html partial has Jason's month selection included.
Initially, only years are displayed unless the path of the current item
contains a year then the months are displayed with links for that year and
month.  In addition, the year and month are context sensitive so for the
current year and month, no link is generated.

To support index files under the /messages path, the Rules file was fixed
to change the routing so that the special name translation to the format
/messages/yyyy/mm/dd/message-title is only done if the filename contains
a date.  Same for the select_messages method in lib/default.rb which will
now only find items containing a date in the path.