Delete HTML and .yaml for 2009.11.08_Daniels_-_ Will_God ...
Originally, the source .mp3 file had a space in the name.  The audio update
process was run which created associtated files also with a space in their
name.

The mp3 file has been renamed so these files must also be deleted.  As the
HTML stub didn't have a transcript, it's easier to just delete and it will
be recreated the next time the audio update process is run.