-Created clickable tabs for the billboard panel and aligned them vertically along the left side of that section.
- Added several more billboard panels which include the latest images from the photo gallery, e.g. Taiwan, Honduras, El Salvador, Panama, etc.  Also added the moving billboards and the images from Vancouver and Vanuatu.

- Added the latest flags/countries to the first panel.

- Added quite a few more May 21 tracts request comments.

- In jquery.innerfade.js file, appended the new case numbers for each panel to keep the fade times accurate.

- Created an image for the verse on the top right of the home page.  Originally tried this with CSS, but since some of the features aren't supported in IE, I decided to do it in Photoshop.  The image is about the size of a billboard image (15K).

- Copied over all of the new/updated images to /var/www/assets.

There's one issue that I am still trying to figure out, but may need some help on.  After the home page first loads and the billboard controls pop up, the control buttons are not clickable.  Also, the tabs on the left are not clickable either.  This only happens for the first two panels (Countries/flags and the first Africa panel).  When the third panel fades in, the buttons become clickable.