Update Rules to handle a new file rewrites.conf(.txt)
To make it easier to manage rewrites with the Nginx web server which does
not support Apache .htaccess files, a rewrites.conf include file is
supported in the nanoc content folder which is configured to be included
within Nginx.

In the content folder, the file is named rewrites.conf.txt and is processed
by the Rules file to be rewrites.conf in the output folder.