Added multilingual study series support
Logic update to support study series in multiple languages. Now, instead
of having to manually enter a unique series id in the "series"
attribute, the value "true" can be used instead and the series id will
be automatically determined by a subset of the URI and all studies in
the same folder will be assumed to be in the same series.

In addition, a "published" attribute can be specified with a value of
"false" which allows staging certain updates without them being visible
when the site is published.

An attribute, "canonical_identifier" is generated (unless specified
explicitly) which will later be used to cross link any given page to all
the same pages in other available languages.

Changes by file:
- Rules
* Updated to support "true" for a series attribute in which case the
base of the path is used as the series id.  Explicit series id's are
also supported for backwards compatibility.
* If a published attribute is set to false, the is_hidden attribute is
set to true.
* A default canonical_identifier created if not specified explicitly.
* Support for multilingual /archives path where English remains at the
top level but other languages use their respective prefix.

- lib/default.rb - Updated method language_code_of(item) to return 'en'
as the default when a path does not start with a two-letter language
code.
- layouts/_items_nav.html - Replaced "Previous" and "Next" with left and
right chevrons respectively to avoid having to translate these words for
each language.

- layouts/default.html - Adjustments to breadcrumbs to not display the
"Home" link unless the page is an English one.  This is required because
English is at the top level, not on the same level with other languages.