Update tasks/deploy_audio.rake to use better defaults for HTML stub files
When an audio file is processed and uploaded, the associated stub file
needs editing for attributes like created_at date, description, etc. This
change sets reasonable defaults for different types of files such as the
Chinese broadcasts, studies and Q&A's so less manual editing is required.

In the case of the Chinese broadcast files, the created_at attribute is
set to 14:00:00 UTC which is the starting time of the broadcast. This
in part prevents files from being displayed on the index page until after
the broadcast starts for that day.